# Replace spacePattern in TextView word wrapping with strings.TrimRightFunc